fix bugs with multiplayer input edit box
